src/main/java/com/priceminister/account/implementation/CustomerAccountRule.java
/*
 * =============================================================================
 *
 *   PRICE MINISTER APPLICATION
 *   Copyright (c) 2000 Babelstore.
 *   All Rights Reserved.
 *
 *   $Source$
 *   $Revision$
 *   $Date$
 *   $Author$
 *
 * =============================================================================
 */
package com.priceminister.account.implementation;

import org.apache.commons.lang3.ObjectUtils;

import com.priceminister.account.AccountRule;


public class CustomerAccountRule implements AccountRule {

    /* (non-Javadoc)
     * @see com.priceminister.account.AccountRule#withdrawPermitted(java.lang.Double)
     */
    public boolean withdrawPermitted(Double resultingAccountBalance) {
        Double amount = ObjectUtils.defaultIfNull(resultingAccountBalance, 0d);
        return amount >= 0;
    }

}
